# config/supla-api-docs.yaml
openapi: 3.0.0
servers:
  - url: 'https://cloud.supla.org/api/v2.3.0'
info:
  version: "2.3.0"
  title: SUPLA Cloud API
paths:
...
  /server-status:
    get:
      tags:
        - Server
      summary: Get the SUPLA Server status
      operationId: getSuplaServerStatus
      security: []
      responses:
        '200':
          description: Supla Server is alive
        '503':
          description: SUPLA Server is down
...
security:
  - BearerAuth: []
  - OAuth2: []
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: 'https://cloud.supla.org/oauth/v2/auth'
          tokenUrl: 'https://cloud.supla.org/oauth/v2/token'
          scopes:
            accessids_r: Access Identifiers (Read)
            accessids_rw: Access Identifiers (Read/Write)
            account_r: User Account (Read)
            account_rw: User Account (Read/Write)
            channels_r: Channels (Read)
            channels_rw: Channels (Read/Write)
            channels_ea: Channels (Execute Actions)
            channels_files: Channels (download files, e.g. icons, measurement logs)
            channelgroups_r: Channel Groups (Read)
            channelgroups_rw: Channel Groups (Read/Write)
            channelgroups_ea: Channel Groups (Execute Actions)
            clientapps_r: Client Apps (Read)
            clientapps_rw: Client Apps (Read/Write)
            directlinks_r: Direct Links (Read)
            directlinks_rw: Direct Links (Read/Write)
            iodevices_r: IO Devices (Read)
            iodevices_rw: IO Devices (Read/Write)
            locations_r: Locations (Read)
            locations_rw: Locations (Read/Write)
            schedules_r: Schedules (Read)
            schedules_rw: Schedules (Read/Write)
            offline_access: Request refresh token